Ignore last commit, FCFS does not work yet.
Write system call goes on forever, so the code has 
been changed back to a state where it works for most
write system call. Need to understand what has caused 
that system call to not work under the new way of
doing this. Simply printing things to the console 
works fine in the new way.