Predict CPU should be useful now.
I made it that PredictCPU yields the CPU after carrying out the syscall
so that all processes in the userThreads list carry out PredictCPU
before any of them actually execute their loop. This required code in
FCFS, the syscall, and UserThread. Thus extra code will be needed for
SJF and any other scheduling policies in the future.